Removed kf library, switched to custom kf

+++ src/robot/src/KalmanFilter.h
@@ -16,25 +16,25 @@ class KalmanFilter
   public:
 
     KalmanFilter(
-        const Eigen::MatrixXd& A,
-        const Eigen::MatrixXd& B,
-        const Eigen::MatrixXd& C,
-        const Eigen::MatrixXd& Q,
-        const Eigen::MatrixXd& R
+        const Eigen::MatrixXf& A,
+        const Eigen::MatrixXf& B,
+        const Eigen::MatrixXf& C,
+        const Eigen::MatrixXf& Q,
+        const Eigen::MatrixXf& R
     );
 
     // Size only simplify initializing in a class
     KalmanFilter(int n, int m);
 
     // Prediction step with input u
-    void predict(const Eigen::VectorXd& u);
+    void predict(const Eigen::VectorXf& u);
 
     // Update the estimated state based on measured values.
-    void update(const Eigen::VectorXd& y);
+    void update(const Eigen::VectorXf& y);
 
     // Get current state estimate
-    Eigen::VectorXd state() { return x_hat_; };
-    Eigen::MatrixXd covariance() { return P_; };
+    Eigen::VectorXf state() { return x_hat_; };
+    Eigen::MatrixXf covariance() { return P_; };
 
   private:
 
@@ -43,18 +43,18 @@ class KalmanFilter
     int m_;
 
     // Matrices for computation
-    Eigen::MatrixXd A_;
-    Eigen::MatrixXd B_;
-    Eigen::MatrixXd C_;
-    Eigen::MatrixXd I_;
-    Eigen::MatrixXd K_;
-    Eigen::MatrixXd P_;
-    Eigen::MatrixXd Q_;
-    Eigen::MatrixXd R_;
-    Eigen::MatrixXd S_;
+    Eigen::MatrixXf A_;
+    Eigen::MatrixXf B_;
+    Eigen::MatrixXf C_;
+    Eigen::MatrixXf I_;
+    Eigen::MatrixXf K_;
+    Eigen::MatrixXf P_;
+    Eigen::MatrixXf Q_;
+    Eigen::MatrixXf R_;
+    Eigen::MatrixXf S_;
 
     // Estimated state
-    Eigen::VectorXd x_hat_;
+    Eigen::VectorXf x_hat_;
 };
 
 #endif //KalmanFilter_h
